💥:  Drop all the support for snowpack
Although it was a very long journey of headaches to setup snowpack, and  even going further to create custom plugins for it, And after finally succeeding in all tasks that I suppose to do, I come to realize that even though it's a very beautiful concept and its work just fine at a very decent point, It's still unpredictable and it can break at any time, every update from the team was introducing new bugs and old bugs where rarely fixed I even try to contribute myself but there was a huge changes going on at the time that always break my work, well Nothing is wasted and it was a great experience but, unfortunately I decide to move to webpack , I start setting webpack up and I was succeeded to make them both work parallel in the same project. But snowpack become a legacy code for me and time has come to drop all packages related to it and remove the config file as it's no longer useful for the project, especially after the dev team move to another path and give-up on the project. Vite the supposedly snowpack alternative was a horrible experience for me as its filled with so much code and functionalities that has a very few uses cases, so from now we are sticking with webpack till the end of this project.

✨ I gain so much knowledge in the last year and I'm exiting to learn more as everyday something new appears✨ .